6.1.4
Used Fetch to grab the api for the repos page.  Then we selected the Network tab in dev tools to see a list of all the resources the browser requested upon loading the page.  We then selected repos from that list because it had the "type" fetch which references what we just put in our JS.   Upon clicking that we selected headers and found the 200 code to show it worked and clicked on response to see what was given back to us.

var getUserRepos = function () {
    fetch("https://api.github.com/users/octocat/repos");
};

getUserRepos();