6.1.5
fetch("https://api.github.com/users/octocat/repos").then(function (response) {
this line grabs the api using fetch and then uses .then(function(response) { to get the data in this response var.

        response.json().then(function (data) {
then we take that response var and use .JSON to format it as a JSON in the data var.
            console.log(data);
we console log the new data var to see the array that has been given.
        });
    });