6.2.4
+var userFormEl = document.querySelector("#user-form");
var nameInputEl = document.querySelector("#username");
these grab the form's id and the input's id and brings them into the HTML and assigns them each a var

var formSubmitHandler = function (event) {
    event.preventDefault();
    starts a function and uses preventDefault so the browser doesn't send the input data to our URL and instead we send it where we want using our following JS

    // get value from input element
    var username = nameInputEl.value.trim();
takes the value of what is in the input field and assigns it a var.  uses trim to get rid of excess spaces for example "Johnny   " makes it look like "Johnny"

    if (username) {
        getUserRepos(username);
if there is a value in the username var it calls the previously created getUserRepos fun and passed it it in (username).

        nameInputEl.value = "";
then it empties the text field

    } else {
        alert("Please enter a GitHub username");
    }
if the text box is empty we get this alert
};

userFormEl.addEventListener("submit", formSubmitHandler);
the call that starts it all.  Whenever anything gets clicked  in this case "submit" (also the enter key to work as well) anywhere in this form (in this case the button el) it starts the formSubmitHandler function just created.