6.3.4
var getRepoIssues = function (repo) {
called function that passes in the var from the call and changes it's name to repo

    var apiUrl = "https://api.github.com/repos/" + repo + "/issues?direction=asc";
assigns that name to this URL that is now targeting the issues page.  We also used the ? to change the direction of the list from descending to ascending and we assign all this to the apiUrl var.

    fetch(apiUrl).then(function (response) {
        // request was successful
        if (response.ok) {
            response.json().then(function (data) {
                console.log(data);
            });
standard fetch response that is checking to see if it worked and logs it
        }
        else {
            alert("There was a problem with your request!");
if not it alerts you
        }
    });

    console.log(repo);
};

getRepoIssues("facebook/react");
the call that starts it all and passes in a string to place in the URL