6.3.5
displayIssues(data);
created a call for this function inside the if statement of the getReposIssues func's fetch "if" the fetch was successful call this new func which....

var displayIssues = function (issues) {
it takes the passed in data and changes it's name to issues

    if (issues.length === 0) {
        issueContainerEl.textContent = "This repo has no open issues!";
        return;
then checks the length of the array given, if 0 it applies the text "This repo has no issues!" and then uses return to get us out of this function.  Otherwise...
    }

    for (var i = 0; i < issues.length; i++) {
this iterates over the issues array

        // create a link element to take users to the issue on github
        var issueEl = document.createElement("a");
creates an a element so it is clickable as a link

        issueEl.classList = "list-item flex-row justify-space-between align-center";
styles it

        issueEl.setAttribute("href", issues[i].html_url);
searches the issues array for the link to the page which in this case is called html_url and gives it the href="linkTextHere" attribute

        issueEl.setAttribute("target", "_blank");
then it gives it the target="_blank" attribute to open link in a new window

        // create span to hold issue title
        var titleEl = document.createElement("span");
        titleEl.textContent = issues[i].title;
next we create a span element element giving it the var titleEl and then set it's textContent to what we find in the issues title in the array which in this case is the issues name.

        // append to container
        issueEl.appendChild(titleEl);

        // create a type element
        var typeEl = document.createElement("span");

        // check if issue is an actual issue or a pull request
        if (issues[i].pull_request) {
            typeEl.textContent = "(Pull request)";
        } else {
            typeEl.textContent = "(Issue)";
        }
this one is confusing because I can't find an example of pull_request in anything.  But if it was to show up this span's text would show up as Pull Request.  But since I have not seen it show up everything just gets added as Issue which is what we are looking for anyways.
        // append to container
        issueEl.appendChild(typeEl);

        issueContainerEl.appendChild(issueEl);
appends it to the HTML
    }
    console.log(issues);

};