6.3.6
if (response.headers.get("Link")) {
                    displayWarning(repo);
                }
in get repos we added an "if" statment within an "if" statment.  The first what that was already there checks "if" there are issues and calls the func to create them.  Then "if" that is true we check "if" the header as something in it called link, (a link is only present if we get more than 30 responses) "if" that is true we run this new function

+var displayWarning = function (repo) {
    // add text to warning container
    limitWarningEl.textContent = "To see more than 30 issues, visit ";

    var linkEl = document.createElement("a");
    linkEl.textContent = "See More Issues on GitHub.com";
    linkEl.setAttribute("href", "https://github.com/" + repo + "/issues");
    linkEl.setAttribute("target", "_blank");

we added a new div to the HTML and appeneded yet another link <a> element to it linking us to the exact repo so they can see all the issues.