6.4.5
moved the newly created added text and func call within the getRepoName func into an "if"  to only run if things are good.

  if (repoName) {
        // display repo name on the page
        repoNameEl.textContent = repoName;

        getRepoIssues(repoName);
    } else {
        // if no repo was given, redirect to the homepage
        document.location.replace("./index.html");
    }
then added an else to return you back to the homepage index.html otherwise.

tweaked the else statment in getRepoIssues to take you back t o the main page as well instead of showing an alert.
 else {
           document.location.replace("./index.html");
        }