6.5.3
This is more important than it seems.  Made a new function that will..

var getFeaturedRepos = function(language) {
  var apiUrl = "https://api.github.com/search/repositories?q=" + language + "+is:featured&sort=help-wanted-issues";
receive a var and call it language and then pass use that to build a URL for the API we need.  The first bit searches the the github api for any repository.  Then we pass in the name of what language we want to search for, in this case javascript was used in the console to test it.  Then oloowing the ?q="our var" we make sure it is featured and sort it by things that have the help-wanted-issues id.

  fetch(apiUrl);
this goes and gets it
};

We used console log to test it by going to the calling it and passing in Javascript and then going to the Network/Preview tabs to get a closer look.