6.5.4
var getFeaturedRepos = function(language) {
  var apiUrl = "https://api.github.com/search/repositories?q=" + language + "+is:featured&sort=help-wanted-issues";

   fetch(apiUrl).then(function(response) {
    if (response.ok) {
      response.json().then(function(data) {
        displayRepos(data.items, language);
      });
    } else {
      alert('Error: GitHub User Not Found');
    }
  });

added a .then  func that created the var response.  Then created an if statment making sure the response var has what was fetched.  If so it converts it to a json and then calls a func that changes it's name to data.  We then call the displayRepos func and pass in data.items because that is the name of the array of 30 and pass in language so we can have that text (in this case the word javascript) to apply to the top of the HTML.  Added the else just in case the fetch failed so we can alert the user.