6.5.6
var languageButtonsEl = document.querySelector("#language-buttons");
selects the div that holds all the buttons and brings it into the JS

var buttonClickHandler = function(event) {
pass in event so I can use .target

  var language = event.target.getAttribute("data-language")
  console.log(language);
created a variable that will target the specific button that is clicked and grab the value of teh data-language atrribute

  if (language) {
    getFeaturedRepos(language);
if it gets a value it pumps that into the func getFeaturedRepos.  The values could be javascript, css and html which getFeaturedRepos can work with.

    // clear old content
    repoContainerEl.textContent = "";
clear old repositories from the main list just in case there are any from previous searches
  }
};

languageButtonsEl.addEventListener("click", buttonClickHandler);
listens for a click anywhere on the div that holds the 3 buttons and calls the buttonClickHandler function once heard.